requirements.txt
openpyxl==3.0.9
python-dotenv==0.19.2
requests==2.26.0
tenacity==8.0.1